+++ app/controllers/roles_controller.rb
@@ -2,4 +2,14 @@
 
 class RolesController < WulinMaster::ScreenController
   controller_for_screen RoleScreen
+
+  private
+
+  def setup_missing_permission
+    if params[:screen].to_s == "AddRoleForUserScreen"
+      create_permission(WulinPermits::UserManagement::SCREEN_CUD)
+    else
+      super
+    end
+  end
 end